tradingview/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
attrs==25.3.0
av==12.3.0
beautifulsoup4==4.13.4
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
curl_cffi==0.11.4
dateparser==1.2.2
django-elasticsearch-dsl==8.0
dnspython==1.16.0
elastic-transport==8.17.1
elasticsearch==8.18.1
elasticsearch-dsl==8.18.0
frozendict==2.4.6
frozenlist==1.7.0
google-crc32c==1.5.0
idna==3.10
ifaddr==0.2.0
multidict==6.6.2
multitasking==0.0.11
numpy==2.3.1
pandas==2.3.0
peewee==3.18.1
platformdirs==4.3.8
propcache==0.3.2
protobuf==6.31.1
pycparser==2.22
pycryptodome==3.23.0
pymongo==3.12.0
python-binance==1.0.29
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2024.11.6
requests==2.32.4
six==1.17.0
soupsieve==2.7
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
websockets==15.0.1
yarl==1.20.1
yfinance==0.2.64
